Refactor: Modular architecture, CI pipeline, copyright update, and gitignore fix
