+++ .gitignore
@@ -105,7 +105,6 @@ dist/
 # Gatsby files
 .cache/
  
-public/
 # Storybook build outputs
 .out
 .storybook-out
