+++ contributors/contributor.html
@@ -130,8 +130,8 @@
     </button>
 
     <!-- Pushed JavaScript Code To Here -->
-    <script src="../index.js"></script>
-    <script src="contributor.js"></script>
+    <script type="module" src="../index.js"></script>
+    <script type="module" src="contributor.js"></script>
     
   </body>
 </html>
