+++ index.html
@@ -170,6 +170,6 @@
     </button>
 
     <script type="module" src="hero-terminal.js"></script>
-    <script src="index.js"></script>
+    <script type="module" src="index.js"></script>
 </body>
 </html>
