+++ modules/config.js
@@ -1,0 +1,6 @@
+export const REPO_OWNER = "dhairyagothi";
+export const REPO_NAME  = "100_days_100_web_project";
+
+// Expose to window for backward compatibility with contributor.js
+window.REPO_OWNER = REPO_OWNER;
+window.REPO_NAME = REPO_NAME;
