+++ modules/theme.js
@@ -1,0 +1,19 @@
+export function initTheme() {
+    const btn = document.getElementById('themeToggle');
+    if (!btn) return;
+
+    const icon = btn.querySelector('i');
+    const saved = localStorage.getItem('theme') || 'dark';
+
+    if (saved === 'light') {
+        document.body.classList.add('light-mode');
+        if (icon) icon.className = 'fas fa-sun';
+    }
+
+    btn.addEventListener('click', () => {
+        document.body.classList.toggle('light-mode');
+        const isLight = document.body.classList.contains('light-mode');
+        if (icon) icon.className = isLight ? 'fas fa-sun' : 'fas fa-moon';
+        localStorage.setItem('theme', isLight ? 'light' : 'dark');
+    });
+}
